RES-303 modified dimensionの数字は"でクオートしない

+++ W2V.hpp
@@ -21,10 +21,10 @@ namespace W2V {
 		long long w2i(const std::string& w);
 		const vector& i2v(int i);
 		long long getNumWords() { return words; }
-		long long getSize() { return size; }
+		int getSize() { return size; }
     private:
 		long long words;
-		long long size;
+		int size;
         std::map<std::string, long long> w2i_;
         std::map<long long, std::string> i2w_;
         std::vector<vector> M;
